Added landing page, updated ActionButton component
