+++ locales/en/common.json
@@ -29,7 +29,7 @@
   "service-and-benefits": "Service and Benefits",
   "skip-to": {
     "about-government": "Skip to 'About government'",
-    "mainContent": "Skip to main content"
+    "main-content": "Skip to main content"
   },
   "title": "home",
   "tools": "Tools",
