+++ locales/fr/home.json
@@ -1,41 +1,7 @@
 {
-  "birth-date": {
-    "error": {
-      "current": "La date de naissance doit être une date dans le passé",
-      "invalid": "La date de naissance doit être une date valide au format (yyyy-MM-dd)",
-      "required": "La date de naissance est obligatoire"
-    },
-    "label": "Date de naissance"
-  },
-  "check-again": "Pour vérifier l'état d'une autre demande, cliquez sur le bouton ci-dessous:",
-  "check-status": "Vérifier l'état",
-  "description": "Utilisez ce service avec votre ESRF pour vérifier le statut de votre demande de passeport.",
-  "esrf": {
-    "error": {
-      "length": "Le numéro de dossier doit comporter 8 caractères",
-      "required": "Le numéro de dossier est obligatoire"
-    },
-    "label": "Numéro de dossier"
-  },
-  "given-name": {
-    "error": {
-      "required": "Le prénom est obligatoire"
-    },
-    "label": "Prénom"
-  },
-  "go-back": "Retourner",
   "header": "Vérification du statut du passeport",
-  "status": {
-    "APPROVED": "Approuvée",
-    "IN_EXAMINATION": "En examen",
-    "REJECTED": "Rejetée"
-  },
-  "status-is": "Le dernier statut de votre candidature selon nos dossiers est : ",
-  "surname": {
-    "error": {
-      "required": "Le nom de famille est obligatoire"
-    },
-    "label": "Nom de famille"
-  },
-  "unable-to-find-status": "Nous ne sommes pas en mesure de déterminer le statut de votre demande pour le moment en utilisant ce service."
+  "description": "Afin de mieux vous servir, veuillez sélectionner l'option ci-dessous qui vous convient le mieux.",
+  "intakeForm": "https://www.canada.ca/fr/immigration-refugies-citoyennete/services/passeports-canadiens/verifier-demande-passeports-documents-voyage.html",
+  "withESRF": "J'ai mon numéro de dossier (ESRF)",
+  "withoutESRF": "Je n'ai pas mon numéro de dossier (ESRF)"
 }
