+++ pages/status.tsx
@@ -1,0 +1,222 @@
+import InputField from '../components/InputField'
+import ActionButton from '../components/ActionButton'
+import { FC, MouseEventHandler, useCallback, useMemo, useState } from 'react'
+import ErrorSummary, {
+  ErrorSummaryItem,
+  getErrorSummaryItem,
+} from '../components/ErrorSummary'
+import Layout from '../components/Layout'
+import { CheckStatusReponse, CheckStatusRequestBody } from './api/check-status'
+import { useCheckStatus } from '../hooks/api/useCheckStatus'
+import { useFormik } from 'formik'
+import * as Yup from 'yup'
+import useTranslation from 'next-translate/useTranslation'
+import Error from 'next/error'
+
+const Status: FC = () => {
+  const { t } = useTranslation('status')
+
+  const [formSubmitted, setFormSubmitted] = useState(false)
+
+  const formik = useFormik<CheckStatusRequestBody>({
+    initialValues: {
+      birthDate: '',
+      esrf: '',
+      givenName: '',
+      surname: '',
+    },
+    validationSchema: Yup.object({
+      esrf: Yup.string()
+        .required('esrf.error.required')
+        .length(8, 'esrf.error.length'),
+      givenName: Yup.string().required('given-name.error.required'),
+      surname: Yup.string().required('surname.error.required'),
+      birthDate: Yup.date()
+        .required('birth-date.error.required')
+        .max(new Date(), 'birth-date.error.current'),
+    }),
+    validateOnBlur: false,
+    validateOnChange: false,
+    validateOnMount: false,
+    onReset: () => {
+      setFormSubmitted(false)
+      removeCheckStatusResponse()
+    },
+    onSubmit: (values) => {
+      setFormSubmitted(true)
+    },
+  })
+
+  const {
+    isLoading: isCheckStatusLoading,
+    error: checkStatusError,
+    data: checkStatusReponse,
+    remove: removeCheckStatusResponse,
+  } = useCheckStatus(formik.values, {
+    enabled: formik.isValid && formSubmitted,
+  })
+
+  const errorSummary = useMemo<ErrorSummaryItem[]>(() => {
+    return Object.keys(formik.errors)
+      .filter((key) => !!formik.errors[key as keyof typeof formik.errors])
+      .map((key) =>
+        getErrorSummaryItem(
+          key,
+          t(formik.errors[key as keyof typeof formik.errors] as string)
+        )
+      )
+  }, [formik, t])
+
+  const handleReset: MouseEventHandler<HTMLButtonElement> = useCallback(
+    (e) => {
+      e.preventDefault()
+      formik.resetForm()
+    },
+    [formik]
+  )
+
+  const handleGoBack: MouseEventHandler<HTMLButtonElement> = useCallback(
+    (e) => {
+      e.preventDefault()
+      setFormSubmitted(false)
+      removeCheckStatusResponse()
+    },
+    [removeCheckStatusResponse]
+  )
+
+  return (
+    <Layout>
+      <h1 className="mb-4">{t('header')}</h1>
+      {checkStatusError ? (
+        <Error statusCode={checkStatusError.statusCode} />
+      ) : checkStatusReponse ? (
+        <PassportStatusInfo
+          checkStatusResponse={checkStatusReponse}
+          handleGoBackClick={handleReset}
+        />
+      ) : checkStatusReponse === null ? (
+        <PassportStatusUnavailable handleGoBackClick={handleGoBack} />
+      ) : (
+        <>
+          <div>
+            <p>{t('description')}</p>
+            {errorSummary.length > 0 && (
+              <ErrorSummary
+                id="error-summary-get-status"
+                summary={t('common:found-errors')}
+                errors={errorSummary}
+              />
+            )}
+            <form onSubmit={formik.handleSubmit} id="form-get-status">
+              <InputField
+                id="esrf"
+                name="esrf"
+                label={t('esrf.label')}
+                onChange={formik.handleChange}
+                value={formik.values.esrf}
+                errorMessage={formik.errors.esrf && t(formik.errors.esrf)}
+                textRequired={t('common:required')}
+                required
+              />
+              <InputField
+                id="givenName"
+                name="givenName"
+                label={t('given-name.label')}
+                onChange={formik.handleChange}
+                value={formik.values.givenName}
+                errorMessage={
+                  formik.errors.givenName && t(formik.errors.givenName)
+                }
+                textRequired={t('common:required')}
+                required
+              />
+              <InputField
+                id="surname"
+                name="surname"
+                label={t('surname.label')}
+                onChange={formik.handleChange}
+                value={formik.values.surname}
+                errorMessage={formik.errors.surname && t(formik.errors.surname)}
+                textRequired={t('common:required')}
+                required
+              />
+              <InputField
+                id="birthDate"
+                name="birthDate"
+                type="date"
+                label={t('birth-date.label')}
+                onChange={formik.handleChange}
+                value={formik.values.birthDate}
+                errorMessage={
+                  formik.errors.birthDate && t(formik.errors.birthDate)
+                }
+                textRequired={t('common:required')}
+                required
+              />
+              <ActionButton
+                disabled={isCheckStatusLoading}
+                type="submit"
+                text={t('check-status')}
+                style="primary"
+              />
+            </form>
+          </div>
+        </>
+      )}
+    </Layout>
+  )
+}
+
+export interface PassportStatusInfoProps {
+  handleGoBackClick: MouseEventHandler<HTMLButtonElement>
+  checkStatusResponse: CheckStatusReponse
+}
+
+export const PassportStatusInfo: FC<PassportStatusInfoProps> = ({
+  checkStatusResponse,
+  handleGoBackClick,
+}) => {
+  const { t } = useTranslation('status')
+  return (
+    <div id="response">
+      <p className="mb-6 text-2xl">
+        {t('status-is')}{' '}
+        <strong>
+          {t(`status.${checkStatusResponse.status}`, null, {
+            default: checkStatusResponse.status,
+          })}
+        </strong>
+        .
+      </p>
+      <p className="mb-6 text-2xl">{t('check-again')}</p>
+      <ActionButton
+        onClick={handleGoBackClick}
+        text={t('go-back')}
+        style="primary"
+      />
+    </div>
+  )
+}
+
+export interface PassportStatusUnavailableProps {
+  handleGoBackClick: MouseEventHandler<HTMLButtonElement>
+}
+
+export const PassportStatusUnavailable: FC<PassportStatusUnavailableProps> = ({
+  handleGoBackClick,
+}) => {
+  const { t } = useTranslation('status')
+  return (
+    <div id="response">
+      <p className=" mb-6 text-2xl">{t('unable-to-find-status')}</p>
+      <p className="mb-6 text-2xl">{t('check-again')}</p>
+      <ActionButton
+        onClick={handleGoBackClick}
+        text={t('go-back')}
+        style="primary"
+      />
+    </div>
+  )
+}
+
+export default Status
